Port network drivers to i386. * EOI problem on slave PIC is fixed. * Changes on other modules are made to fix some compiling errors.
