+++ ucore/src/kern-ucore/arch/arm/include/memlayout_common.h
@@ -47,18 +47,6 @@ struct memmap {
 
 struct Page;
 
-#ifdef UCONFIG_BIONIC_LIBC
-struct mapped_addr {
-	struct Page *page;
-	off_t offset;
-	list_entry_t list;
-};
-
-#define le2maddr(le)	\
-		to_struct((le), struct mapped_addr, list)
-
-#endif //UCONFIG_BIONIC_LIBC
-
 /* *
  * struct Page - Page descriptor structures. Each Page describes one
  * physical page (1MB). In kern/mm/pmm.h, you can find lots of useful functions
