+++ ucore/src/kern-ucore/arch/i386/driver/Makefile
@@ -1,1 +1,2 @@
-obj-y := clock.o console.o ide.o intr.o picirq.o pci.o
+obj-y := clock.o console.o ide.o intr.o picirq.o pci.o cmos_rtc.o
+dirs-y := network
