+++ ucore/src/kern-ucore/arch/i386/driver/cmos_rtc.h
@@ -1,0 +1,1 @@
+#include "../../amd64/driver/cmos_rtc.h"
