+++ ucore/src/kern-ucore/arch/i386/driver/network/Makefile
@@ -1,0 +1,1 @@
+obj-y := e1000.o
