+++ ucore/src/kern-ucore/arch/i386/driver/network/e1000.c
@@ -1,0 +1,1 @@
+#include "../../../amd64/driver/network/e1000.c"
